fix mappings and repository query requests for new functions (event and registration on event)

+++ src/main/java/ru/trofimov/eventmanager/model/EventFilter.java
@@ -15,7 +15,7 @@ public record EventFilter(
         BigDecimal costMax,
         Integer durationMin,
         Integer durationMax,
-        Integer locationId,
+        Long locationId,
         EventStatus status
 ) {
 }
